Desactiva el producto
desactiva los productos que no puede eli minar

+++ src/page/producto/modificarProducto/ModificarProducto.jsx
@@ -7,11 +7,18 @@ import BtnGuardEditElim from "../../../components/btnCrud/BtnGuardEditElim";
 import { IoCloseSharp } from "react-icons/io5";
 import { validFormProduct } from "../../../validation/formProducto/formProductoVal";
 import { notError, notSuccess } from "../../../components/alert/alert";
+import Swal from "sweetalert2";
 
 export default function ModificarProducto({ closed }) {
   const nameForm = "Modificar Producto";
 
-  const { dataProductModifi, idProductModifi, updateProduct, deleteProduct } = useProductos();
+  const {
+    dataProductModifi,
+    idProductModifi,
+    updateProduct,
+    deleteProduct,
+    desactivateProduct,
+  } = useProductos();
 
   const [nombreProd, setNombreProd] = useState("");
   const [precioProd, setPrecioProd] = useState("");
@@ -20,7 +27,8 @@ export default function ModificarProducto({ closed }) {
   const [codBarProd, setCodBarProd] = useState("");
   const [descripcionProd, setDescripcionProd] = useState("");
   const [img, setImg] = useState(null);
-  const [urlImg, setUrlImg] = useState("")
+  const [urlImg, setUrlImg] = useState("");
+  const [dependencia, setDependencia] = useState("");
 
   const [disabledInput, setDisabledInput] = useState(true);
 
@@ -32,6 +40,7 @@ export default function ModificarProducto({ closed }) {
     setCodBarProd(dataProductModifi.cod_barra || "");
     setDescripcionProd(dataProductModifi.descripcion_pro || "");
     setUrlImg(dataProductModifi.img_prod || "");
+    setDependencia(dataProductModifi.dependencia || "");
   }, [dataProductModifi]);
 
   const formItemsProduc = [
@@ -146,13 +155,15 @@ export default function ModificarProducto({ closed }) {
       formData.append("codBarra", codBarProd);
       formData.append("descripcion", descripcionProd);
       formData.append("id", idProductModifi);
-      formData.append("urlImg", urlImg)
+      formData.append("urlImg", urlImg);
+      formData.append("dependencia", dependencia);
 
       try {
         const response = await updateProduct({ body: formData });
         if (response) {
           notSuccess("Producto Modificado");
-          closed(false)
+          closed(false);
+  
         }
       } catch (error) {
         console.error("Error al crear producto:", error);
@@ -162,19 +173,80 @@ export default function ModificarProducto({ closed }) {
     }
   };
 
-  const eliminarProducto = async ()=>{
+  const eliminarProducto = async () => {
     try {
-      const resultado = await deleteProduct(idProductModifi);
+      Swal.fire({
+        title: "Desea eliminar este Producto",
+        showDenyButton: true,
+        showCancelButton: true,
+        confirmButtonText: "Si",
+        denyButtonText: `no`,
+        confirmButtonColor: "#29C716",
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          const resultado = await deleteProduct(idProductModifi);
 
-      if(resultado.error){
-        return resultado.error;
-      }else{
-        return resultado;
-      }
+          if (resultado === "Producto eliminado") {
+            Swal.fire({
+              title: "Producto Eliminado Correctamente",
+              icon: "success",
+              confirmButtonColor: "#29C716",
+            }).then((result)=>{
+              if(result.isConfirmed){
+                window.location.reload();
+              }
+            });
+
+            closed(false);
+          } else {
+            if (resultado != "") {
+              Swal.fire({
+                icon: "error",
+                title: "¿Desea solo desactivar el producto?",
+                text: resultado.error,
+                showDenyButton: true,
+                showCancelButton: true,
+                confirmButtonText: "Si",
+                denyButtonText: `no`,
+                confirmButtonColor: "#29C716",
+              }).then(async (result) => {
+                if (result.isConfirmed) {
+                  const res = await desactivateProduct(idProductModifi);
+
+                  if(res.status == 200){
+                    Swal.fire({
+                      title: "Producto desactivado Correctamente",
+                      icon: "success",
+                      confirmButtonColor: "#29C716",
+                    }).then((result)=>{
+                      if (result.isConfirmed){
+                        window.location.reload();
+                        closed(!closed)
+                      }
+                    });
+                  }
+                } else if (result.isDenied) {
+                  Swal.fire({
+                    title: "No se desactivo el Producto",
+                    icon: "info",
+                    confirmButtonColor: "#29C716",
+                  });
+                }
+              });
+            }
+          }
+        } else if (result.isDenied) {
+          Swal.fire({
+            title: "No se elimino ningun Producto",
+            icon: "info",
+            confirmButtonColor: "#29C716",
+          });
+        }
+      });
     } catch (error) {
       console.error("Error al ejecutar la promesa:", error);
     }
-  }
+  };
 
   return (
     <div className="container-modificar-producto">
@@ -192,7 +264,7 @@ export default function ModificarProducto({ closed }) {
           enableInput={setDisabledInput}
           closed={closed}
           saved={updateProducto}
-          eliminar = {eliminarProducto}
+          eliminar={eliminarProducto}
         />
       </div>
     </div>
